Make batch enhancer more general
