+++ devtools/client/webconsole/new-console-output/actions/moz.build
@@ -4,6 +4,7 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 DevToolsModules(
+    'enhancers.js',
     'filters.js',
     'index.js',
     'messages.js',
